Cambios Especialidad y ObraSocial

+++ src/basic/views/especialidad/index.php
@@ -2,62 +2,61 @@
 /* @var $this yii\web\View */
 
 //$this->registerCssFile("https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/css/bootstrap.min.css",['position'=>$this::POS_HEAD]);
-$this->registerCssFile("//unpkg.com/bootstrap/dist/css/bootstrap.min.css",['position'=>$this::POS_HEAD]);
-$this->registerCssFile("//unpkg.com/bootstrap-vue@latest/dist/bootstrap-vue.min.css",['position'=>$this::POS_HEAD]);
+$this->registerCssFile("//unpkg.com/bootstrap/dist/css/bootstrap.min.css", ['position' => $this::POS_HEAD]);
+$this->registerCssFile("//unpkg.com/bootstrap-vue@latest/dist/bootstrap-vue.min.css", ['position' => $this::POS_HEAD]);
 
-$this->registerJsFile("https://cdn.jsdelivr.net/npm/vue/dist/vue.js",['position'=>$this::POS_HEAD]);
-$this->registerJsFile("https://unpkg.com/bootstrap-vue@latest/dist/bootstrap-vue.min.js",['position'=>$this::POS_HEAD]);
+$this->registerJsFile("https://cdn.jsdelivr.net/npm/vue/dist/vue.js", ['position' => $this::POS_HEAD]);
+$this->registerJsFile("https://unpkg.com/bootstrap-vue@latest/dist/bootstrap-vue.min.js", ['position' => $this::POS_HEAD]);
 
-$this->registerJsFile("https://unpkg.com/bootstrap-vue@latest/dist/bootstrap-vue-icons.min.js",['position'=>$this::POS_HEAD]);
+$this->registerJsFile("https://unpkg.com/bootstrap-vue@latest/dist/bootstrap-vue-icons.min.js", ['position' => $this::POS_HEAD]);
 
-$this->registerJsFile("https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js",['position'=>$this::POS_HEAD]);
+$this->registerJsFile("https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js", ['position' => $this::POS_HEAD]);
+$this->registerJsFile("https://cdn.jsdelivr.net/npm/sweetalert2@9", ['position' => $this::POS_HEAD]);
 
 ?>
 
 <div id="app" class="container">
-    <h1>{{msg}}</h1>
+    <template>
+        <h1>{{msg}}
+            <b-icon icon="card-checklist"></b-icon>
+        </h1>
+    </template>
     <!-- Button trigger modal -->
-        <b-modal v-model="showModal" id="my-modal">
-                <form action="">
-                    <div class="form-group">
-                        <label for="nombre">Nombre</label>
-                        <input v-model="especialidad.nombre" type="text" name="nombre" id="nombre" class="form-control" placeholder="Ingrese especialidad" aria-describedby="helpId">
-                        <small id="titlehelpId" class="text-muted"></small>
-                        <span class="text-danger" v-if="errors.nombre"> {{ errors.nombre }} </span>
-                    </div>
-                    <div class="form-group">
-                        <label for="detalle">Descripcion</label>
-                        <input v-model="especialidad.detalle" type="text" name="detalle" id="detalle" class="form-control" placeholder="Ingrese descripcion" aria-describedby="helpId">
-                        <small id="bodyhelpId" class="text-muted"></small>
-                        <span class="text-danger" v-if="errors.detalle">{{ errors.detalle }}</span>
-                    </div>
-                    
-                </form>
-                <template v-slot:modal-footer="{ok, cancel, hide}">
-                    <button v-if="isNewRecord"  @click="addEspecialidad()" type="button" class="btn btn-primary m-3">Crear</button>
-                    <button v-if="!isNewRecord" @click="isNewRecord = !isNewRecord" v-on:click="especialidad={}" type="button" class="btn btn-success m-3">Nuevo</button>
-                    <button v-if="!isNewRecord" @click="updateEspecialidad(especialidad.id_especialidad)" type="button" class="btn btn-primary m-3">Actualizar</button>
+    <b-modal v-model="showModal" id="my-modal">
+        <form action="">
+            <div class="form-group">
+                <label for="nombre">Nombre</label>
+                <input v-model="especialidad.nombre" type="text" name="nombre" id="nombre" class="form-control" placeholder="Ingrese especialidad" aria-describedby="helpId">
+                <small id="titlehelpId" class="text-muted"></small>
+                <span class="text-danger" v-if="errors.nombre"> {{ errors.nombre }} </span>
+            </div>
+            <div class="form-group">
+                <label for="detalle">Detalle</label>
+                <input v-model="especialidad.detalle" type="text" name="detalle" id="detalle" class="form-control" placeholder="Ingrese descripcion" aria-describedby="helpId">
+                <small id="bodyhelpId" class="text-muted"></small>
+                <span class="text-danger" v-if="errors.detalle">{{ errors.detalle }}</span>
+            </div>
 
-                </template>      
-        </b-modal>
+        </form>
+        <template v-slot:modal-footer="{ok, cancel, hide}">
+            <button v-if="isNewRecord" @click="addEspecialidad()" type="button" class="btn btn-primary m-3">Crear</button>
+            <!-- <button v-if="!isNewRecord" @click="isNewRecord = !isNewRecord" v-on:click="especialidad={}" type="button" class="btn btn-success m-3">Nuevo</button> -->
+            <button v-if="!isNewRecord" @click="updateEspecialidad(especialidad.id_especialidad)" type="button" class="btn btn-primary m-3">Actualizar</button>
 
-        <b-pagination
-            v-model="currentPage"
-            :total-rows="pagination.total"
-            :per-page="pagination.perPage"
-            aria-controls="my-table"
-        ></b-pagination>
+        </template>
+    </b-modal>
 
-        <p>
-            <button @click="showModal=true" type="button" class="btn btn-primary">Nuevo</button>
-        </p>
 
-        <table class="table">
-            <thead>
+    <p>
+        <button @click="showModal=true" type="button" class="btn btn-primary">Nuevo</button>
+    </p>
+
+    <table class="table">
+        <thead>
             <tr>
                 <th>Id</th>
                 <th>Nombre</th>
-                <th>Descripcion</th>
+                <th>Detalle</th>
                 <th></th>
                 <th></th>
             </tr>
@@ -74,14 +73,14 @@ $this->registerJsFile("https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js",['p
                 <td></td>
                 <td></td>
             </tr>
-            </thead>
+        </thead>
 
-            <tbody>
+        <tbody>
             <tr v-for="(espec,key) in especialidades" v-bind:key="espec.id_especialidad">
                 <td scope="row">{{espec.id_especialidad}}</td>
                 <td>{{espec.nombre}}</td>
                 <td>{{espec.detalle}}</td>
-                
+
                 <td>
                     <button @click="showModal=true" v-on:click="editEspecialidad(key)" type="button" class="btn btn-warning">Editar</button>
                 </td>
@@ -89,50 +88,53 @@ $this->registerJsFile("https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js",['p
                     <button v-on:click="deleteEspecialidad(espec.id_especialidad)" type="button" class="btn btn-danger">Borrar</button>
                 </td>
             </tr>
-            </tbody>
-        </table>
+        </tbody>
+    </table>
+    <b-pagination v-model="currentPage" :total-rows="pagination.total" :per-page="pagination.perPage" aria-controls="my-table"></b-pagination>
 
 </div>
 
 <script>
-
     var app = new Vue({
 
         el: "#app",
-        data: function () {
+        data: function() {
             return {
                 msg: "Especialidades",
                 especialidades: [],
-                especialidad:{},
-                filter:{},
-                errors:{},
-                isNewRecord:true,
+                especialidad: {},
+                filter: {},
+                errors: {},
+                isNewRecord: true,
                 currentPage: 1,
-                pagination:{},
+                pagination: {},
                 showModal: false,
             }
         },
         mounted() {
             this.getEspecialidades();
         },
-        watch:{
-            currentPage: function(){
+        watch: {
+            currentPage: function() {
                 this.getPatologias();
             }
         },
         methods: {
-            normalizeErrors: function(errors){
+            normalizeErrors: function(errors) {
                 var allErrors = {};
-                for(var i = 0 ; i < errors.length; i++){
+                for (var i = 0; i < errors.length; i++) {
                     allErrors[errors[i].field] = errors[i].message;
                 }
                 return allErrors;
 
             },
-            getEspecialidades: function(){
+
+            getEspecialidades: function() {
                 var self = this;
-                axios.get('/apiv1/especialidads?page='+self.currentPage,{params:self.filter})
-                    .then(function (response) {
+                axios.get('/apiv1/especialidads?page=' + self.currentPage, {
+                        params: self.filter
+                    })
+                    .then(function(response) {
                         // handle success
                         console.log(response.data);
                         console.log("traje todas las especialidades");
@@ -141,82 +143,102 @@ $this->registerJsFile("https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js",['p
                         self.pagination.perPage = parseInt(response.headers['x-pagination-per-page']);
                         self.especialidades = response.data;
                     })
-                    .catch(function (error) {
+                    .catch(function(error) {
                         // handle error
                         self.errors = self.normalizeErrors(error.response.data);
                         console.log(self.errors);
                     })
-                    .then(function () {
+                    .then(function() {
                         // always executed
                     });
             },
-            deleteEspecialidad: function(id){
-                var self = this;
-                axios.delete('/apiv1/especialidads/'+ id)
-                    .then(function (response) {
-                        // handle success
-                        console.log("borra especialidad id: "+ id);
-                        console.log(response.data);
-                        self.getEspecialidades();
-                    })
-                    .catch(function (error) {
-                        // handle error
-                        console.log(error);
-                    })
-                    .then(function () {
-                        // always executed
-                    });
+            deleteEspecialidad: function(id) {
+                Swal.fire({
+                    title: 'Esta seguro que desea borrar el registro ' + id + '?',
+                    icon: 'warning',
+                    showCancelButton: true,
+                    showConfirmButton: true,
+                    confirmButtonText: 'Si borrar!',
+                    cancelButtonText: 'No, regresar.',
+                }).then((result) => {
+                    if (result.value) {
+                        var self = this;
+                        axios.delete('/apiv1/especialidads/' + id)
+                            .then(function(response) {
+                                // handle success
+                                console.log("borra especialidad id: " + id);
+                                console.log(response.data);
+                                self.getEspecialidades();
+                            })
+                            .catch(function(error) {
+                                // handle error
+                                console.log(error);
+                            })
+                            .then(function() {
+                                // always executed
+                            });
+                        Swal.fire({
+                            title: 'Se ha borrado con exito',
+                            icon: 'success',
+                        })
+                    }
+                }, );
+
+
+
+                // if (this.showConfirmButton = true) {
             },
-            editEspecialidad: function (key) {
-                this.especialidad = Object.assign({},this.especialidades[key]);
+
+            editEspecialidad: function(key) {
+                this.especialidad = Object.assign({}, this.especialidades[key]);
                 this.especialidad.key = key;
                 this.isNewRecord = false;
             },
-            addEspecialidad: function(){
+            addEspecialidad: function() {
                 var self = this;
-                axios.post('/apiv1/especialidads',self.especialidad)
-                    .then(function (response) {
+                axios.post('/apiv1/especialidads', self.especialidad)
+                    .then(function(response) {
                         // handle success
                         console.log(response.data);
                         self.getEspecialidades()
                         // self.posts.unshift(response.data);
                         self.especialidad = {};
+                        self.showModal = false;
                     })
-                    .catch(function (error) {
+                    .catch(function(error) {
                         // handle error
                         self.errors = self.normalizeErrors(error.response.data);
                         console.log(self.errors);
 
                     })
-                    .then(function () {
+                    .then(function() {
                         // always executed
                     });
             },
-            updateEspecialidad: function (key) {
+            updateEspecialidad: function(key) {
                 var self = this;
                 const params = new URLSearchParams();
                 params.append('nombre', self.especialidad.nombre);
                 params.append('detalle', self.especialidad.detalle);
-                axios.patch('/apiv1/especialidads/'+key,params)
-                    .then(function (response) {
+                axios.patch('/apiv1/especialidads/' + key, params)
+                    .then(function(response) {
                         // handle success
                         console.log(response.data);
                         self.getEspecialidades();
                         self.especialidad = {};
                         self.isNewRecord = true;
+                        self.showModal = false;
                     })
-                    .catch(function (error) {
+                    .catch(function(error) {
                         // handle error
                         console.log(error);
                     })
-                    .then(function () {
+                    .then(function() {
                         // always executed
                     });
-
             }
 
         }
 
     })
-
 </script>
